Move build-bergamot.py into tree as build-wasm.py
Some of the functionality that was previously defined in build-bergamot.py
within Mozilla Central is now fit to live directly within this repository.

https://searchfox.org/mozilla-central/rev/dca2603d55b5b39d3b8ab8e93c08b42563f5aad8/toolkit/components/translations/bergamot-translator/build-bergamot.py

This commit simply copies the file directly over as build-wasm.py so that
the next commit can be more clear about how the file is modified from its
original source.