Off-ramp
This is all fucked up. Dataclass has some magic to be supported, and I
can't get the type system and dynamic checks to agree on the primitives.
I took a detour to look at NamedTuple. It can't be subclassed multiple
times as far as I can tell.

It might be worth looking into protocols to solve things. I'm unsure how
those types work. Or it might be worth to just have a static class that
handles the serialization and deserialization strategies to cut the
gordian knot of the Python type system.

- Experiment with Protocols
- Cut the Gordian Knot